Move condition fields from options tab to the specification tab
