+++ src/PrestaShopBundle/Form/Admin/Sell/Product/Specification/SpecificationsType.php
@@ -28,13 +28,36 @@ declare(strict_types=1);
 
 namespace PrestaShopBundle\Form\Admin\Sell\Product\Specification;
 
+use PrestaShop\PrestaShop\Core\Form\FormChoiceProviderInterface;
 use PrestaShopBundle\Form\Admin\Sell\Product\Options\ProductAttachmentsType;
+use PrestaShopBundle\Form\Admin\Type\SwitchType;
 use PrestaShopBundle\Form\Admin\Type\TranslatorAwareType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Translation\TranslatorInterface;
 
 class SpecificationsType extends TranslatorAwareType
 {
+    /**
+     * @var FormChoiceProviderInterface
+     */
+    private $productConditionChoiceProvider;
+
+    /**
+     * @param TranslatorInterface $translator
+     * @param array $locales
+     * @param FormChoiceProviderInterface $productConditionChoiceProvider
+     */
+    public function __construct(
+        TranslatorInterface $translator,
+        array $locales,
+        FormChoiceProviderInterface $productConditionChoiceProvider
+    ) {
+        parent::__construct($translator, $locales);
+        $this->productConditionChoiceProvider = $productConditionChoiceProvider;
+    }
+
     /**
      * {@inheritDoc}
      */
@@ -44,6 +67,22 @@ class SpecificationsType extends TranslatorAwareType
             ->add('references', ReferencesType::class)
             ->add('features', FeaturesType::class)
             ->add('attachments', ProductAttachmentsType::class)
+            ->add('condition', ChoiceType::class, [
+                'choices' => $this->productConditionChoiceProvider->getChoices(),
+                'attr' => [
+                    'class' => 'custom-select',
+                ],
+                'required' => false,
+                // placeholder false is important to avoid empty option in select input despite required being false
+                'placeholder' => false,
+                'label' => $this->trans('Condition', 'Admin.Catalog.Feature'),
+                'label_tag_name' => 'h2',
+                'label_help_box' => $this->trans('Not all shops sell new products. This option enables you to indicate the condition of the product. It can be required on some marketplaces.', 'Admin.Catalog.Help'),
+            ])
+            ->add('show_condition', SwitchType::class, [
+                'required' => false,
+                'label' => $this->trans('Display condition on product page', 'Admin.Catalog.Feature'),
+            ])
             ->add('customizations', CustomizationsType::class)
         ;
     }
